stackfusion-assignment/backend-django/requirements.txt
Django>=5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
